pyproject.toml
[tool.poetry]
name = "python-prompting"
version = "0.1.0"
description = ""
authors = ["alextanhongpin <alextan220990@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.2.2"
streamlit = "^1.28.2"
black = "^23.11.0"
pandas = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
